pyproject.toml
[project]
name = "studyflow-backend"
version = "0.1.0"
description = "Backend for StudyFlow"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.4",
    "fastapi-mail>=1.6.1",
    "mangum>=0.19.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.124.4",
    "uvicorn>=0.38.0",
]
